add employee avater
