+++ client/app/app/PanelInbox/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 import * as React from 'react'
 import { Tabs, TabsList, TabsTrigger, TabsContent } from '@radix-ui/react-tabs'
 import { Input } from 'components/ui/input'
@@ -5,18 +6,21 @@ import { ResizablePanel, ResizableHandle } from 'components/ui/resizable'
 import { Separator } from 'components/ui/separator'
 import { Search } from 'lucide-react'
 import { MailList } from '../components/mail-list'
-import { mails } from '../data'
+import { mails } from './data'
 import { MailDisplay } from '../components/mail-display'
 import { useMail } from '../use-mail'
+import { AppStores } from 'lib/zustand'
 
 export default function PanelInbox() {
   const [mail] = useMail()
+  const store = AppStores.useEmployee()
+
   return (
     <>
-      <ResizablePanel defaultSize={30} minSize={30}>
+      <ResizablePanel defaultSize={30} minSize={30} maxSize={35}>
         <Tabs defaultValue="all">
-          <div className="flex items-center px-4 py-2 h-[60px]">
-            <h1 className="text-xl font-bold">Inbox</h1>
+          <div className="flex h-[60px] items-center px-4 py-2">
+            <h1 className="text-xl font-bold">Employees</h1>
             <TabsList className="ml-auto">
               <TabsTrigger value="all" className="mx-2 text-zinc-600 dark:text-zinc-200">
                 All mail
@@ -44,8 +48,8 @@ export default function PanelInbox() {
         </Tabs>
       </ResizablePanel>
       <ResizableHandle withHandle />
-      <ResizablePanel defaultSize={50} minSize={30}>
-        <MailDisplay mail={mails.find((item) => item.id === mail.selected) || null} />
+      <ResizablePanel defaultSize={70} minSize={40}>
+        <MailDisplay mail={mails.filter((item) => item.key === store.active)[0] || null} />
       </ResizablePanel>
     </>
   )
