+++ client/app/app/components/mail.tsx
@@ -36,30 +36,8 @@ export function Mail({
           settingsStore.update({ defaultLayout: sizes })
           document.cookie = `react-resizable-panels:layout:mail=${JSON.stringify(sizes)}`
         }}
-        className="h-full max-h-[800px] items-stretch"
+        className="h-full items-stretch"
       >
-        {/* <ResizablePanel
-          defaultSize={defaultLayout[0]}
-          collapsedSize={navCollapsedSize}
-          collapsible={true}
-          minSize={15}
-          maxSize={20}
-          onCollapse={() => {
-            setIsCollapsed(true)
-            settingsStore.update({ defaultCollapsed: true })
-            // document.cookie = `react-resizable-panels:collapsed=${JSON.stringify(true)}`
-          }}
-          onResize={() => {
-            setIsCollapsed(false)
-            settingsStore.update({ defaultCollapsed: false })
-            // document.cookie = `react-resizable-panels:collapsed=${JSON.stringify(false)}`
-          }}
-          className={cn(isCollapsed && 'min-w-[50px] transition-all duration-300 ease-in-out')}
-        >
-          <Separator />
-          <Nav isCollapsed={isCollapsed} links={navData} />
-          <Separator />
-        </ResizablePanel> */}
         <PanelInbox />
       </ResizablePanelGroup>
     </TooltipProvider>
