+++ client/components/ui/separator.tsx
@@ -13,7 +13,7 @@ const Separator = React.forwardRef<
     orientation={orientation}
     className={cn(
       'shrink-0 bg-border',
-      orientation === 'horizontal' ? 'h-[1px] w-full' : 'h-full w-[1px]',
+      orientation === 'horizontal' ? 'h-[0.2px] w-full' : 'h-full w-[0.2px]',
       className,
     )}
     {...props}
