+++ client/lib/zustand/index.ts
@@ -1,7 +1,9 @@
+import { useEmployee } from './employee'
 import { useView } from './screens'
 import { useSettings } from './settings'
 
 export const AppStores = {
   useSettings,
   useView,
+  useEmployee,
 }
